+++ src/gex/commands/file/file.py
@@ -1,6 +1,7 @@
 import click
 
 from .deinterleave import deinterleave
+from .hash import hash
 from .identify import identify
 from .slice import slice
 
@@ -10,5 +11,6 @@ def file():
     pass
 
 file.add_command(deinterleave)
+file.add_command(hash)
 file.add_command(identify)
 file.add_command(slice)