+++ src/gex/commands/postprocess/steam/list.py
@@ -1,8 +1,4 @@
 import click
-import os
-import glob
-import importlib
-import pkgutil
 from gex.lib.transforms import *
 import inspect
 
@@ -12,5 +8,6 @@ def list():
     for global_name, transform_module in globals().items():
         if inspect.ismodule(transform_module) and transform_module.__package__ == 'gex.lib.transforms':
             print(f'{global_name}: {transform_module.title}')
-            print(f'  {transform_module.description}')
-            print(f'  Expected input dir: {transform_module.in_dir_desc}')
+            if len(transform_module.description) > 0:
+                print(f'  {transform_module.description}')
+            print(f'  Expected input dir: {transform_module.in_dir_desc} (ex. {transform_module.default_folder})')