+++ src/gex/lib/transforms/cas1_old.py
@@ -5,9 +5,10 @@ import logging
 import os
 import io
 
-
 from gex.lib.archive import kpka
-from gex.lib.utils import blob
+from gex.lib.utils.blob import transforms
+
+logger = logging.getLogger('gextoolbox')
 
 # Standard processing:
 # - Assume each zip in the KPKA is a game
@@ -15,7 +16,8 @@ from gex.lib.utils import blob
 
 title = "Capcom Arcade Stadium 1 (OLD)"
 description = "Capcom Arcade Stadium 1 as downloaded from depot using old manifests"
-in_dir_desc = "SteamApps Content Folder (ex. C:\Program Files (x86)\Steam\steamapps\content)"
+default_folder = "C:\Program Files (x86)\Steam\steamapps\content"
+in_dir_desc = "SteamApps Content Folder"
 
 pkg_name_map = {
     "1515951": "1943",
@@ -54,20 +56,20 @@ pkg_name_map = {
 
 def debug_print_kpka_contents(kpka_contents):
     for key, entry in kpka_contents.items():
-        logging.debug(f'{key}: offset {entry["offset"]}, size {entry["size"]}')
-        logging.debug(f'    {entry["contents"][0:30]}')
+        logger.debug(f'{key}: offset {entry["offset"]}, size {entry["size"]}')
+        logger.debug(f'    {entry["contents"][0:30]}')
 
 def debug_print_zip_contents(zip_bytes):
     with zipfile.ZipFile(io.BytesIO(zip_bytes), "r") as old_archive:
         zip_entries = list(old_archive.infolist())
         for file_entry in zip_entries:
-            logging.debug(f'{file_entry.filename}')
+            logger.debug(f'{file_entry.filename}')
 
 def twiddle_zip(zip_bytes, remove_list = [], rename_dict = {}, lowercase_all = False):
     new_contents = io.BytesIO()
     with zipfile.ZipFile(io.BytesIO(zip_bytes), "r") as old_archive:
         zip_entries = list(old_archive.infolist())
-        with zipfile.ZipFile(new_contents, "w") as new_archive:
+        with zipfile.ZipFile(new_contents, "w", compression=zipfile.ZIP_DEFLATED) as new_archive:
             for file_entry in zip_entries:
                 # Skip files to remove
                 if not file_entry.filename in remove_list:
@@ -100,7 +102,7 @@ def merged_rom_handler(zip_contents, func_map):
                     new_data.update(type_func(file_data))
     # Build the new zip file
     new_contents = io.BytesIO()
-    with zipfile.ZipFile(new_contents, "w") as new_archive:
+    with zipfile.ZipFile(new_contents, "w", compression=zipfile.ZIP_DEFLATED) as new_archive:
         for name, data in new_data.items():
             new_archive.writestr(name, data)
     return new_contents.getvalue()
@@ -275,8 +277,8 @@ def handle_package_1556718(kpka_contents):
 def handle_package_1556722(kpka_contents):
     out_files = []
     def gfx(contents):
-        chunks = blob.custom_split(contents, [8388608, 4194304, 4194304])
-        chunks = blob.deinterleave_all(chunks, num_ways=4, word_size=2)
+        chunks = transforms.custom_split(contents, [8388608, 4194304, 4194304])
+        chunks = transforms.deinterleave_all(chunks, num_ways=4, word_size=2)
         filenames = [
             'sfx.13m',
             'sfx.15m',
@@ -294,17 +296,17 @@ def handle_package_1556722(kpka_contents):
         return dict(zip(filenames, chunks))
 
     def audiocpu(contents):
-        contents = blob.splice_out(contents, 0x48000, length=0x8000)
-        contents = blob.splice_out(contents, 0x8000, length=0x8000)
-        chunks = blob.equal_split(contents, num_chunks=2)
+        contents = transforms.splice_out(contents, 0x48000, length=0x8000)
+        contents = transforms.splice_out(contents, 0x8000, length=0x8000)
+        chunks = transforms.equal_split(contents, num_chunks=2)
         filenames = [
             'sfx.01',
             'sfx.02'
         ]
         return dict(zip(filenames, chunks))
     def qsound(contents):
-        chunks = blob.equal_split(contents, num_chunks=2)
-        chunks = blob.swap_endian_all(chunks)
+        chunks = transforms.equal_split(contents, num_chunks=2)
+        chunks = transforms.swap_endian_all(chunks)
         filenames = [
             'sfx.11m',
             'sfx.12m'
@@ -317,9 +319,9 @@ def handle_package_1556722(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x380000, length=0x3FFFFF)
-                    chunks = blob.equal_split(contents, num_chunks=7)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x380000, length=0x3FFFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=7)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "sfxj.03d", 
                         "sfxj.04a", 
@@ -339,9 +341,9 @@ def handle_package_1556722(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x380000, length=0x3FFFFF)
-                    chunks = blob.equal_split(contents, num_chunks=7)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x380000, length=0x3FFFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=7)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "sfxu.03e", 
                         "sfxu.04a", 
@@ -362,8 +364,8 @@ def handle_package_1556722(kpka_contents):
 def handle_package_1556724(kpka_contents):
     out_files = []
     def gfx(contents):
-        chunks = blob.equal_split(contents, num_chunks=2)
-        chunks = blob.deinterleave_all(chunks, num_ways=4, word_size=2)
+        chunks = transforms.equal_split(contents, num_chunks=2)
+        chunks = transforms.deinterleave_all(chunks, num_ways=4, word_size=2)
         filenames = [   
             "cyb.13m", 
             "cyb.15m", 
@@ -377,9 +379,9 @@ def handle_package_1556724(kpka_contents):
         return dict(zip(filenames, chunks))
 
     def audiocpu(contents):
-        contents = blob.splice_out(contents, 0x48000, length=0x8000)
-        contents = blob.splice_out(contents, 0x8000, length=0x8000)
-        chunks = blob.equal_split(contents, num_chunks=2)
+        contents = transforms.splice_out(contents, 0x48000, length=0x8000)
+        contents = transforms.splice_out(contents, 0x8000, length=0x8000)
+        chunks = transforms.equal_split(contents, num_chunks=2)
 
         return {
             'cyb.01': chunks[0],
@@ -387,8 +389,8 @@ def handle_package_1556724(kpka_contents):
         }
 
     def qsound(contents):
-        chunks = blob.equal_split(contents, num_chunks=2)
-        chunks = blob.swap_endian_all(chunks)
+        chunks = transforms.equal_split(contents, num_chunks=2)
+        chunks = transforms.swap_endian_all(chunks)
 
         return {
             'cyb.11m': chunks[0],
@@ -401,8 +403,8 @@ def handle_package_1556724(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    chunks = blob.equal_split(contents, num_chunks=8)
-                    chunks = blob.swap_endian_all(chunks)
+                    chunks = transforms.equal_split(contents, num_chunks=8)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "cybj.03", 
                         "cybj.04", 
@@ -423,8 +425,8 @@ def handle_package_1556724(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    chunks = blob.equal_split(contents, num_chunks=8)
-                    chunks = blob.swap_endian_all(chunks)
+                    chunks = transforms.equal_split(contents, num_chunks=8)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "cybu.03", 
                         "cybu.04", 
@@ -451,9 +453,9 @@ def handle_package_1556725(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x280000, length=0x180000)
-                    chunks = blob.equal_split(contents, num_chunks=5)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x280000, length=0x180000)
+                    chunks = transforms.equal_split(contents, num_chunks=5)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [    
                         '19xj.03a',
                         '19xj.04a',
@@ -465,9 +467,9 @@ def handle_package_1556725(kpka_contents):
                 func_map['maincpu'] = maincpu
 
                 def gfx(contents):
-                    contents = blob.splice_out(contents, 0x200000, length=0x600000)
-                    chunks = blob.equal_split(contents, num_chunks=5)
-                    chunks = blob.deinterleave_all(chunks, num_ways=4, word_size=2)
+                    contents = transforms.splice_out(contents, 0x200000, length=0x600000)
+                    chunks = transforms.equal_split(contents, num_chunks=5)
+                    chunks = transforms.deinterleave_all(chunks, num_ways=4, word_size=2)
                     filenames = [    
                         '19x-69.4j',
                         '19x-59.4d',
@@ -495,16 +497,16 @@ def handle_package_1556725(kpka_contents):
                 func_map['gfx'] = gfx
 
                 def audiocpu(contents):
-                    contents = blob.splice_out(contents, 0x28000, length=0x28000)
-                    contents = blob.splice_out(contents, 0x8000, length=0x8000)
+                    contents = transforms.splice_out(contents, 0x28000, length=0x28000)
+                    contents = transforms.splice_out(contents, 0x8000, length=0x8000)
                     return {
                         '19x-01.1a': contents
                     }
                 func_map['audiocpu'] = audiocpu
 
                 def qsound(contents):
-                    chunks = blob.equal_split(contents, num_chunks=8)
-                    chunks = blob.swap_endian_all(chunks)
+                    chunks = transforms.equal_split(contents, num_chunks=8)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "19x-51.6a", 
                         "19x-52.7a", 
@@ -524,9 +526,9 @@ def handle_package_1556725(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x280000, length=0x180000)
-                    chunks = blob.equal_split(contents, num_chunks=5)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x280000, length=0x180000)
+                    chunks = transforms.equal_split(contents, num_chunks=5)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [    
                         '19xu.03',
                         '19xu.04',
@@ -538,9 +540,9 @@ def handle_package_1556725(kpka_contents):
                 func_map['maincpu'] = maincpu
                 
                 def gfx(contents):
-                    contents = blob.splice_out(contents, 0x200000, length=0x600000)
-                    chunks = blob.custom_split(contents, [2097152, 8388608])
-                    chunks = blob.deinterleave_all(chunks, num_ways=4, word_size=2)
+                    contents = transforms.splice_out(contents, 0x200000, length=0x600000)
+                    chunks = transforms.custom_split(contents, [2097152, 8388608])
+                    chunks = transforms.deinterleave_all(chunks, num_ways=4, word_size=2)
                     filenames = [    
                         "19x.13m", 
                         "19x.15m", 
@@ -555,16 +557,16 @@ def handle_package_1556725(kpka_contents):
                 func_map['gfx'] = gfx
                 
                 def audiocpu(contents):
-                    contents = blob.splice_out(contents, 0x28000, length=0x28000)
-                    contents = blob.splice_out(contents, 0x8000, length=0x8000)
+                    contents = transforms.splice_out(contents, 0x28000, length=0x28000)
+                    contents = transforms.splice_out(contents, 0x8000, length=0x8000)
                     return {
                         '19x.01': contents
                     }
                 func_map['audiocpu'] = audiocpu
 
                 def qsound(contents):
-                    chunks = blob.equal_split(contents, num_chunks=2)
-                    chunks = blob.swap_endian_all(chunks)
+                    chunks = transforms.equal_split(contents, num_chunks=2)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "19x.11m",   
                         "19x.12m"
@@ -578,7 +580,7 @@ def handle_package_1556726(kpka_contents):
     out_files = []
 
     def gfx(contents):
-        chunks = blob.deinterleave(contents, num_ways=4, word_size=2)
+        chunks = transforms.deinterleave(contents, num_ways=4, word_size=2)
         filenames = [
             'btc.13m',
             'btc.15m',
@@ -588,17 +590,17 @@ def handle_package_1556726(kpka_contents):
         return dict(zip(filenames, chunks))
 
     def audiocpu(contents):
-        contents = blob.splice_out(contents, 0x48000, length=0x8000)
-        contents = blob.splice_out(contents, 0x8000, length=0x8000)
-        chunks = blob.equal_split(contents, num_chunks=2)
+        contents = transforms.splice_out(contents, 0x48000, length=0x8000)
+        contents = transforms.splice_out(contents, 0x8000, length=0x8000)
+        chunks = transforms.equal_split(contents, num_chunks=2)
         filenames = [
             'btc.01',
             'btc.02'
         ]
         return dict(zip(filenames, chunks))
     def qsound(contents):
-        chunks = blob.equal_split(contents, num_chunks=2)
-        chunks = blob.swap_endian_all(chunks)
+        chunks = transforms.equal_split(contents, num_chunks=2)
+        chunks = transforms.swap_endian_all(chunks)
         filenames = [
             'btc.11m',
             'btc.12m'
@@ -611,9 +613,9 @@ def handle_package_1556726(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x380000, length=0x3FFFFF)
-                    chunks = blob.equal_split(contents, num_chunks=7)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x380000, length=0x3FFFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=7)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "btcj.03", 
                         "btcj.04", 
@@ -633,9 +635,9 @@ def handle_package_1556726(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x380000, length=0x3FFFFF)
-                    chunks = blob.equal_split(contents, num_chunks=7)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x380000, length=0x3FFFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=7)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "btce.03", 
                         "btce.04", 
@@ -657,7 +659,7 @@ def handle_package_1556727(kpka_contents):
     out_files = []
 
     def gfx(contents):
-        chunks = blob.deinterleave(contents, num_ways=4, word_size=2)
+        chunks = transforms.deinterleave(contents, num_ways=4, word_size=2)
         filenames = [
             'ggw.13m',
             'ggw.15m',
@@ -667,12 +669,12 @@ def handle_package_1556727(kpka_contents):
         return dict(zip(filenames, chunks))
 
     def audiocpu(contents):
-        contents = blob.splice_out(contents, 0x28000, length=0x28000)
-        contents = blob.splice_out(contents, 0x8000, length=0x8000)
+        contents = transforms.splice_out(contents, 0x28000, length=0x28000)
+        contents = transforms.splice_out(contents, 0x8000, length=0x8000)
         return {'ggw.01': contents}
     def qsound(contents):
-        chunks = blob.equal_split(contents, num_chunks=2)
-        chunks = blob.swap_endian_all(chunks)
+        chunks = transforms.equal_split(contents, num_chunks=2)
+        chunks = transforms.swap_endian_all(chunks)
         filenames = [
             'ggw.11m',
             'ggw.12m'
@@ -685,9 +687,9 @@ def handle_package_1556727(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x180000, length=0x27FFFF)
-                    chunks = blob.equal_split(contents, num_chunks=3)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x180000, length=0x27FFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=3)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "ggwj.03a",
                         "ggwj.04a",
@@ -703,9 +705,9 @@ def handle_package_1556727(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x180000, length=0x27FFFF)
-                    chunks = blob.equal_split(contents, num_chunks=3)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x180000, length=0x27FFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=3)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "ggwu.03",
                         "ggwu.04",
@@ -723,8 +725,8 @@ def handle_package_1556728(kpka_contents):
     out_files = []
 
     def gfx(contents):
-        chunks = blob.custom_split(contents, [16777216, 4194304])
-        chunks = blob.deinterleave_all(chunks, num_ways=4, word_size=2)
+        chunks = transforms.custom_split(contents, [16777216, 4194304])
+        chunks = transforms.deinterleave_all(chunks, num_ways=4, word_size=2)
         filenames = [    
             "nff.13m", 
             "nff.15m", 
@@ -738,12 +740,12 @@ def handle_package_1556728(kpka_contents):
         return dict(zip(filenames, chunks))
 
     def audiocpu(contents):
-        contents = blob.splice_out(contents, 0x28000, length=0x28000)
-        contents = blob.splice_out(contents, 0x8000, length=0x8000)
+        contents = transforms.splice_out(contents, 0x28000, length=0x28000)
+        contents = transforms.splice_out(contents, 0x8000, length=0x8000)
         return {'nff.01': contents}
     def qsound(contents):
-        chunks = blob.equal_split(contents, num_chunks=2)
-        chunks = blob.swap_endian_all(chunks)
+        chunks = transforms.equal_split(contents, num_chunks=2)
+        chunks = transforms.swap_endian_all(chunks)
         filenames = [
             'nff.11m',
             'nff.12m'
@@ -756,9 +758,9 @@ def handle_package_1556728(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x180000, length=0x27FFFF)
-                    chunks = blob.equal_split(contents, num_chunks=3)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x180000, length=0x27FFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=3)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "nffj.03",
                         "nffj.04",
@@ -774,9 +776,9 @@ def handle_package_1556728(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x180000, length=0x27FFFF)
-                    chunks = blob.equal_split(contents, num_chunks=3)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x180000, length=0x27FFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=3)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "nffu.03",
                         "nff.04",
@@ -794,7 +796,7 @@ def handle_package_1556729(kpka_contents):
     out_files = []
 
     def gfx(contents):
-        chunks = blob.deinterleave(contents, num_ways=8, word_size=1)
+        chunks = transforms.deinterleave(contents, num_ways=8, word_size=1)
         filenames = [    
             "pga-simm.01c",
             "pga-simm.01d",
@@ -808,12 +810,12 @@ def handle_package_1556729(kpka_contents):
         return dict(zip(filenames, chunks))
 
     def audiocpu(contents):
-        contents = blob.splice_out(contents, 0x28000, length=0x28000)
-        contents = blob.splice_out(contents, 0x8000, length=0x8000)
+        contents = transforms.splice_out(contents, 0x28000, length=0x28000)
+        contents = transforms.splice_out(contents, 0x8000, length=0x8000)
         return {'pga.01': contents}
     def qsound(contents):
-        chunks = blob.equal_split(contents, num_chunks=4)
-        chunks = blob.swap_endian_all(chunks)
+        chunks = transforms.equal_split(contents, num_chunks=4)
+        chunks = transforms.swap_endian_all(chunks)
         filenames = [
             "pga-simm.05a",
             "pga-simm.05b",
@@ -828,9 +830,9 @@ def handle_package_1556729(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x100000, length=0x2FFFFF)
-                    chunks = blob.equal_split(contents, num_chunks=2)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x100000, length=0x2FFFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=2)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "pgaj.03",
                         "pgaj.04"
@@ -845,9 +847,9 @@ def handle_package_1556729(kpka_contents):
                 func_map = {}
 
                 def maincpu(contents):
-                    contents = blob.splice_out(contents, 0x100000, length=0x2FFFFF)
-                    chunks = blob.equal_split(contents, num_chunks=2)
-                    chunks = blob.swap_endian_all(chunks)
+                    contents = transforms.splice_out(contents, 0x100000, length=0x2FFFFF)
+                    chunks = transforms.equal_split(contents, num_chunks=2)
+                    chunks = transforms.swap_endian_all(chunks)
                     filenames = [   
                         "pgau.03",
                         "pgau.04"
@@ -878,8 +880,8 @@ def rebuild_mame_subfolder_zip(contents):
         try:
             index = zip_entries[0].filename.index('/')
         except Exception as e:
-            logging.warning(e)
-            logging.warning(zip_entries[0])
+            logger.warning(e)
+            logger.warning(zip_entries[0])
             raise Exception(f'not a mame subfolder zip - no slash in first zip entry')
 
         prefix = getPrefix(zip_entries[0])
@@ -888,7 +890,7 @@ def rebuild_mame_subfolder_zip(contents):
                 raise Exception(f'not a mame subfolder zip - {getPrefix(file_entry)} != {prefix}')
 
         new_contents = io.BytesIO()
-        with zipfile.ZipFile(new_contents, "w") as new_archive:
+        with zipfile.ZipFile(new_contents, "w", compression=zipfile.ZIP_DEFLATED) as new_archive:
             for file_entry in zip_entries:
                 with old_archive.open(file_entry) as file_read_obj:
                     file_data = file_read_obj.read()
@@ -918,28 +920,31 @@ def main(steam_dir, out_path):
         else:
             id = file[-11:-4]
 
-        logging.info(f"Extracting {file}: {pkg_name_map[id]}") 
-        try:
-            with open(file, "rb") as curr_file:
-                file_content = bytearray(curr_file.read())
-                kpka_contents = kpka.extract(file_content)
-                output_files = []
+        if id in pkg_name_map:
+            logger.info(f"Extracting {file}: {pkg_name_map[id]}") 
+            try:
+                with open(file, "rb") as curr_file:
+                    file_content = bytearray(curr_file.read())
+                    kpka_contents = kpka.extract(file_content)
+                    output_files = []
 
-                special_func = globals().get(f'handle_package_{id}')
-                if special_func:
-                    # Reflectively call the appropriate function to process the file
-                    output_files = special_func(kpka_contents)
-                else:
-                    output_files = standard_kpka_contents_processing(kpka_contents)
-                    
-                for output_file in output_files:
-                    with open(os.path.join(out_path, output_file['filename']), "wb") as out_file:
-                        out_file.write(output_file['contents'])
-        except Exception as e:
-            traceback.print_exc()
-            logging.warning(f'Error while processing {file}!') 
+                    special_func = globals().get(f'handle_package_{id}')
+                    if special_func:
+                        # Reflectively call the appropriate function to process the file
+                        output_files = special_func(kpka_contents)
+                    else:
+                        output_files = standard_kpka_contents_processing(kpka_contents)
+                        
+                    for output_file in output_files:
+                        with open(os.path.join(out_path, output_file['filename']), "wb") as out_file:
+                            out_file.write(output_file['contents'])
+            except Exception as e:
+                traceback.print_exc()
+                logger.warning(f'Error while processing {file}!') 
+        else:
+            logger.info(f'Skipping {file} as it contains no known ROMS!') 
 
-    logging.info("""
+    logger.info("""
         Processing complete. 
         TODOs:
          - Figure out fixes for the 'incomplete' games